Support copying to clipboard on Wayland
Implement automatic copying of TOTP codes to clipboard on Wayland
sessions.
Detection of Wayland vs X11 is done with the `XDG_SESSION_TYPE`
environment variable and defaults to X11 (previous behaviour) if unset.
The copy is performed using the `wl-copy` command from the
`wl-clipboard` package, which is the same dependency that `pass` uses.
The clipboard type (`primary`/`clipboard`) is tweakable like for X11,
with the same environment variable, and defaults to `clipboard` like for
X11.